Fix intro preloader asset base path handling
